+++ src/main/java/domain/card/Deck.java
@@ -1,8 +1,4 @@
-package domain;
-
-import domain.card.Card;
-import domain.card.CardNumber;
-import domain.card.CardShape;
+package domain.card;
 
 import java.util.*;
 import java.util.stream.Collectors;
