+++ src/main/java/domain/participant/Dealer.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.participant;
 
+import domain.card.Deck;
 import domain.card.Card;
 
 import java.util.List;
