+++ src/main/java/dto/GameScoreDTO.java
@@ -1,7 +1,7 @@
 package dto;
 
-import domain.Dealer;
-import domain.Player;
+import domain.participant.Dealer;
+import domain.participant.Player;
 
 import java.util.ArrayList;
 import java.util.List;
