+++ src/test/java/domain/participant/DealerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import java.util.List;
