+++ src/test/java/domain/participant/PlayerTest.java
@@ -1,8 +1,10 @@
-package domain;
+package domain.participant;
 
+import domain.GameResult;
 import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardShape;
+import domain.card.Deck;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
